requirements.txt
Flask==2.0.3
numpy==1.21.4
pandas==1.3.4
python-dotenv==0.19.2
requests==2.11.1
gunicorn